examples/quickstart/requirements.txt
pytidb[models]
python-dotenv 